+++ product/src/main/java/com/i4u/product/infrastructure/ProductRepositoryImpl.java
@@ -4,11 +4,11 @@ package com.i4u.product.infrastructure;
 import com.i4u.product.application.dto.ProductSearchCond;
 import com.i4u.product.domain.Product;
 import com.i4u.product.domain.QProduct;
-import com.i4u.product.domain.repository.ProductJpaRepository;
 import com.i4u.product.domain.repository.ProductRepository;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.*;
 import org.springframework.stereotype.Repository;
 
@@ -16,91 +16,90 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@Repository
+//@Repository
 //@RequiredArgsConstructor
-public class ProductRepositoryImpl implements ProductRepository {
-
-    private ProductJpaRepository productJpaRepository;
-    private JPAQueryFactory queryFactory;
-    private final QProduct product = QProduct.product;
-
-    //상품 저장
-    @Override
-    public Product save(Product product) {
-        return productJpaRepository.save(product);
-    }
-
-    //쿼리 DSL 사용하기 위한 코드
-    public ProductRepositoryImpl(JPAQueryFactory queryFactory) {
-        this.queryFactory = queryFactory;
-    }
-
-    //쿼리 DSL - 상품 아이디로 Product찾기
-    @Override
-    public Optional<Product> findById(UUID id) {
-        Product result = queryFactory
-                .selectFrom(product)
-                .where(product.id.eq(id))
-                .fetchOne();
-        return Optional.ofNullable(result);
-    }
-
-    //상품 전체 조회 쿼리DSL
-    @Override
-    public Page<Product> findAll(Pageable pageable) {
-        List<Product> products = queryFactory
-                .selectFrom(product)
-                .where(product.deletedAt.isNull()) // deleted_at이 null인 데이터만 조회
-                .offset(pageable.getOffset())
-                .limit(pageable.getPageSize())
-                .fetch();
-
-        long total = queryFactory   //전체 데이터 개수
-                .select(product.count())        //전체 개수 조회
-                .from(product)
-                .where(product.deletedAt.isNull()) // deleted_at이 null인 데이터만 조회
-                .fetchFirst();  //fetchOne 대신 fetchFirst 사용 -> null이어도 예외 발생하지 않게
-
-        //products : 현재 페이지 상품 목록, pageable 페이징 정보, total 전체상품개수 - deletedAt null인 상품만
-        return new PageImpl<>(products, pageable, total);
-    }
-
-    //조건에 맞는 상품 조회
-    @Override
-    public Page<Product> search(ProductSearchCond cond, Pageable pageable) {
-        BooleanBuilder builder = getBooleanBuilder(cond);
-        List<Product> products = queryFactory
-                .selectFrom(product)
-                .where(builder)
-                .offset(pageable.getOffset())
-                .limit(pageable.getPageSize())
-                .fetch();
-
-        long total = queryFactory
-                .select(product.count())
-                .from(product)
-                .where(builder)
-                .fetchOne();
-        return new PageImpl<>(products, pageable, total);
-    }
-
-    private BooleanBuilder getBooleanBuilder(ProductSearchCond cond) {
-        BooleanBuilder builder = new BooleanBuilder();
-
-//        if (cond.hubId() != null) {
-//            builder.and(product.hub.id.eq(cond.hubId()));
+public abstract class ProductRepositoryImpl implements ProductRepository {
+//    private final ProductRepository productRepository;
+//    private final JPAQueryFactory queryFactory;
+//    private final QProduct product = QProduct.product;
+//
+//    //상품 저장
+//    @Override
+//    public Product save(Product product) {
+//        return productRepository.save(product);
+//    }
+//
+//    //쿼리 DSL 사용하기 위한 코드
+//    public ProductRepositoryImpl(JPAQueryFactory queryFactory, ProductRepository productRepository) {
+//        this.queryFactory = queryFactory;
+//        this.productRepository = productRepository;
+//    }
+//
+//    //쿼리 DSL - 상품 아이디로 Product찾기
+//    @Override
+//    public Optional<Product> findById(UUID id) {
+//        Product result = queryFactory
+//                .selectFrom(product)
+//                .where(product.id.eq(id))
+//                .fetchOne();
+//        return Optional.ofNullable(result);
+//    }
+//
+//    //상품 전체 조회 쿼리DSL
+//    @Override
+//    public Page<Product> findAll(Pageable pageable) {
+//        List<Product> products = queryFactory
+//                .selectFrom(product)
+//                .where(product.deletedAt.isNull()) // deleted_at이 null인 데이터만 조회
+//                .offset(pageable.getOffset())
+//                .limit(pageable.getPageSize())
+//                .fetch();
+//
+//        long total = queryFactory   //전체 데이터 개수
+//                .select(product.count())        //전체 개수 조회
+//                .from(product)
+//                .where(product.deletedAt.isNull()) // deleted_at이 null인 데이터만 조회
+//                .fetchFirst();  //fetchOne 대신 fetchFirst 사용 -> null이어도 예외 발생하지 않게
+//
+//        //products : 현재 페이지 상품 목록, pageable 페이징 정보, total 전체상품개수 - deletedAt null인 상품만
+//        return new PageImpl<>(products, pageable, total);
+//    }
+//
+//    //조건에 맞는 상품 조회
+//    public Page<Product> search(ProductSearchCond cond, Pageable pageable) {
+//        BooleanBuilder builder = getBooleanBuilder(cond);
+//        List<Product> products = queryFactory
+//                .selectFrom(product)
+//                .where(builder)
+//                .offset(pageable.getOffset())
+//                .limit(pageable.getPageSize())
+//                .fetch();
+//
+//        long total = queryFactory
+//                .select(product.count())
+//                .from(product)
+//                .where(builder)
+//                .fetchOne();
+//        return new PageImpl<>(products, pageable, total);
+//    }
+//
+//    private BooleanBuilder getBooleanBuilder(ProductSearchCond cond) {
+//        BooleanBuilder builder = new BooleanBuilder();
+//
+////        if (cond.hubId() != null) {
+////            builder.and(product.hub.id.eq(cond.hubId()));
+////        }
+////        if (cond.companyId() != null) {
+////            builder.and(product.company.id.eq(cond.companyId()));
+////       }
+//        if (cond.name() != null) {
+//            builder.and(product.name.containsIgnoreCase(cond.name()));
 //        }
-//        if (cond.companyId() != null) {
-//            builder.and(product.company.id.eq(cond.companyId()));
-//       }
-        if (cond.name() != null) {
-            builder.and(product.name.containsIgnoreCase(cond.name()));
-        }
-        if (!cond.isDeleted()) {
-            builder.and(product.isDeleted.eq(false));
-        }
-        return builder;
-    }
+//        if (!cond.isDeleted()) {
+//            builder.and(product.isDeleted.eq(false));
+//        }
+//        return builder;
+//    }
 
 //    private List<Product> fetchProducts(ProductSearchCond cond, BooleanBuilder builder,
 //                                        Pageable pageable) {
